support urls in ffs stage command
this makes `pow ffs stage https://url-to-file` work, expanding the potential sources for
staging into powergate. I've smoke-tested with a 250Mb file stored in an S3 bucket.

Urls *only* write to files (no folder support), which I think covers most url-oriented
use-cases.

One caveat: if a user has a file that starts with 'http', it'll be a problem. The easiest
fix is to refer to the file with a './' prefix, such as: './http'

Signed-off-by: b5 <sparkle_pony_2000@qri.io>